Moved Fusion context endpoint
