+++ backend/src/Equinor.ProjectExecutionPortal.WebApi/Controllers/OnboardedContextController.cs
@@ -57,20 +57,5 @@ namespace Equinor.ProjectExecutionPortal.WebApi.Controllers
 
             return NoContent();
         }
-
-        [HttpGet("fusion/{externalId}")]
-        public async Task<ActionResult<FusionContext>> GetFusionContext(string externalId)
-        {
-            // External id eg: FC5FFCBC-392F-4D7E-BB14-79A006579337
-            var contextIdentifier = ContextIdentifier.FromExternalId(externalId);
-            var context = await ContextResolver.ResolveContextAsync(contextIdentifier, FusionContextType.ProjectMaster);
-
-            if (context == null)
-            {
-                return FusionApiError.NotFound(externalId, "Could not find context by external id");
-            }
-
-            return context;
-        }
     }
 }
